refactor(AllRubrosMaterialsTable): move component to its own directory
